use aurora client specific env vars & add rds client
