+++ .env
@@ -1,4 +1,4 @@
 LOCAL_METAB_PORT=8080
-METADB_CLUSTER_ARN=arn:aws:rds:us-east-1:123456789012:cluster:dummy
-METADB_SECRET_ARN=arn:aws:secretsmanager:us-east-1:123456789012:secret:dummy
+AURORA_CLUSTER_ARN=arn:aws:rds:us-east-1:123456789012:cluster:dummy
+AURORA_SECRET_ARN=arn:aws:secretsmanager:us-east-1:123456789012:secret:dummy
 METADB_NAME=postgres
