+++ docker/src/create_deploy_stack/create_deploy_stack.py
@@ -25,6 +25,9 @@ log.setLevel(logging.DEBUG)
 
 ssm = boto3.client("ssm")
 lb = boto3.client("lambda")
+rds_data_client = boto3.client(
+    "rds-data", endpoint_url=os.environ.get("METAB_LOCAL_ENDPOINT")
+)
 
 
 class CreateStack:
@@ -226,6 +229,7 @@ class CreateStack:
             aurora_cluster_arn=os.environ["METADB_CLUSTER_ARN"],
             secret_arn=os.environ["METADB_SECRET_ARN"],
             database=os.environ["METADB_NAME"],
+            rds_data_client=rds_data_client,
         ) as conn:
             with conn.cursor() as cur:
                 cur.execute(
