+++ tests/unit/functions/approval_response/test_app.py
@@ -5,7 +5,8 @@ import sys
 import json
 from unittest.mock import patch
 from pprint import pformat
-from tests.helpers.utils import insert_records, local_conn
+from tests.helpers.utils import insert_records, rds_data_client
+import aurora_data_api
 from functions.approval_response.lambda_function import App, ClientException
 from contextlib import nullcontext as does_not_raise
 
@@ -81,7 +82,7 @@ execution_id = "run-123"
     os.environ,
     {"METADB_CLUSTER_ARN": "mock", "METADB_SECRET_ARN": "mock", "METADB_NAME": "mock"},
 )
-@pytest.mark.usefixtures("mock_conn", "aws_credentials", "truncate_executions")
+@pytest.mark.usefixtures("aws_credentials", "truncate_executions")
 @patch("boto3.client")
 def test_update_vote(
     mock_boto_client,
@@ -110,7 +111,9 @@ def test_update_vote(
 
     if record != {}:
         log.info("Assert record's approriate voters list contains the recipient")
-        with local_conn() as conn, conn.cursor() as cur:
+        with aurora_data_api.connect(
+            database=os.environ["METADB_NAME"], rds_data_client=rds_data_client
+        ) as conn, conn.cursor() as cur:
             cur.execute(
                 """
                     SELECT approval_voters, rejection_voters
